Renamed module and fixed test package
