+++ fabric-dehardcoding-api-v0/build.gradle
@@ -1,4 +1,4 @@
-archivesBaseName = "fabric-dehardcoding-api-v0"
+archivesBaseName = "fabric-extensibility-api-v1"
 version = getSubprojectVersion(project, "1.0.0")
 
 dependencies {
