+++ fabric-dehardcoding-api-v0/src/main/resources/fabric-extensibility-api-v1.mixins.json
