+++ fabric-dehardcoding-api-v0/src/main/resources/fabric.mod.json
@@ -1,33 +1,28 @@
 {
   "schemaVersion": 1,
-  "id": "fabric-dehardcoding-api-v0",
+  "id": "fabric-extensibility-api-v1",
+  "name": "Fabric Extensibility API (v1)",
   "version": "${version}",
-
-  "name": "Fabric Dehardcoding API (v0)",
-  "description": "This is an example description! Tell everyone what your mod is about!",
-  "authors": [
-    "FabricMC"
-  ],
+  "environment": "*",
+  "license": "Apache-2.0",
+  "icon": "assets/fabric-extensibility-api-v1/icon.png",
   "contact": {
     "homepage": "https://fabricmc.net",
     "irc": "irc://irc.esper.net:6667/fabric",
     "issues": "https://github.com/FabricMC/fabric/issues",
     "sources": "https://github.com/FabricMC/fabric"
   },
-  "license": "Apache-2.0",
-  
-  "icon": "assets/fabric-dehardcoding-api-v0/icon.png",
-  
-  "mixins": [
-    "fabric-dehardcoding-api-v0.mixins.json"
+  "authors": [
+    "FabricMC"
   ],
-
+  "mixins": [
+    "fabric-extensibility-api-v1.mixins.json"
+  ],
   "depends": {
-    "fabricloader": ">=0.7.4",
-    "fabric": "*",
-    "minecraft": "1.16.x"
+    "fabricloader": ">=0.4.0",
+    "fabric-api-base": "*"
   },
-  "suggests": {
-    "flamingo": "*"
-  }
+  "description": "Makes the hardcoded features of Minecraft more extendable"
 }
+
+
