+++ fabric-dehardcoding-api-v0/src/testmod/java/net/fabricmc/fabric/test/extensibility/FabricCrossbowTests.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.test.tool;
+package net.fabricmc.fabric.test.extensibility;
 
 import net.minecraft.item.ArrowItem;
 import net.minecraft.item.CrossbowItem;
