+++ settings.gradle
@@ -19,7 +19,7 @@ include 'fabric-blockrenderlayer-v1'
 include 'fabric-commands-v0'
 include 'fabric-command-api-v1'
 include 'fabric-containers-v0'
-include 'fabric-dehardcoding-api-v0'
+include 'fabric-extensibility-api-v1'
 include 'fabric-content-registries-v0'
 include 'fabric-crash-report-info-v1'
 //include 'fabric-dimensions-v1'
